Fix pkcs7 padding - remove changes added for test

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to standardize the PKCS7 padding implementation.

#### PR Summary
This pull request corrects the PKCS7 padding in the encryption workflow by removing a conditional check that was added for testing purposes, ensuring that padding is always applied as per the PKCS7 specification.
- Changes in `/MsvmPkg/AziHsmDxe/AziHsmDxe.c`: Removed the conditional check and unconditionally compute the padding value.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #60134173

+++ MsvmPkg/AziHsmDxe/AziHsmBKS3.h
@@ -77,7 +77,7 @@
 // Forward declared opaque types for TPM sealing outputs.
 typedef struct {
   UINT8     Data[AZIHSM_BUFFER_MAX_SIZE]; // Buffer
-  UINT32    Size;                         // Size of the buffer
+  UINT16    Size;                         // Size of the buffer
 } AZIHSM_BUFFER;
 
 //
